Fix issue #1 by being more robust with incomplete Zotero record
Prevously, the code assumed that a Zotero record would have a
'parentItem' key in the 'data' dictionary.  Issue #1 shows this is
apparently not always the case.  The new code tests for this.  It also
bubbles up errors to main_body, to reduce the number of alert()'s all
over the place.